LanggraphProject/visual-rag/requirements.txt
langchain
langgraph
google-generativeai
gradio
pillow
python-dotenv
